Refactor methods in to classes
